+++ lib/op25_repeater/lib/p25p1_voice_decode.h
@@ -42,7 +42,7 @@ namespace gr {
       // Nothing to declare in this block.
 
      public:
-      p25p1_voice_decode(bool verbose_flag, const char* udp_host, int udp_port, std::deque<int16_t> &_output_queue);
+      p25p1_voice_decode(bool verbose_flag, const char* udp_host, int udp_port, std::deque<float> &_output_queue);
       ~p25p1_voice_decode();
 	void rxframe(const uint32_t u[]);
 	void rxchar(const char* c, int len);
@@ -62,7 +62,7 @@ namespace gr {
 	software_imbe_decoder software_decoder;
 	bool d_software_imbe_decoder;
 
-	std::deque<int16_t> &output_queue;
+	std::deque<float> &output_queue;
 
 	bool opt_verbose;
 	int opt_udp_port;
